refactor: Use composition to decide when to fetch the user profile
