+++ dotcom-rendering/src/web/components/Discussion.stories.tsx
@@ -1,6 +1,6 @@
 import { ArticleDesign, ArticleDisplay, ArticlePillar } from '@guardian/libs';
 
-import { Discussion } from './Discussion.importable';
+import { Discussion } from './Discussion';
 
 export default {
 	component: Discussion,
